Fix esp32 build

+++ include/gpio.h
@@ -47,7 +47,7 @@ static inline bool gpio_read(int pin) {
 #define GPIO_ENABLE1_REG REG(0X3ff4402c)           // Pins 32-39
 
 static inline void gpio_output_enable(int pin, bool enable) {
-  volatile unsigned long *r = GPIO_ENABLE_REG;
+  volatile uint32_t *r = GPIO_ENABLE_REG;
   if (pin > 31) pin -= 31, r = GPIO_ENABLE1_REG;
   r[0] &= ~BIT(pin);
   r[0] |= (enable ? 1U : 0U) << pin;
@@ -59,14 +59,14 @@ static inline void gpio_output(int pin) {
 }
 
 static inline void gpio_write(int pin, bool value) {
-  volatile unsigned long *r = GPIO_OUT_REG;
+  volatile uint32_t *r = GPIO_OUT_REG;
   if (pin > 31) pin -= 31, r = GPIO_OUT1_REG;
-  r[0] &= ~BIT(pin);               // Clear first
+  r[0] &= ~BIT(pin);                 // Clear first
   r[0] |= (value ? 1U : 0U) << pin;  // Then set
 }
 
 static inline void gpio_toggle(int pin) {
-  volatile unsigned long *r = GPIO_OUT_REG;
+  volatile uint32_t *r = GPIO_OUT_REG;
   if (pin > 31) pin -= 31, r = GPIO_OUT1_REG;
   r[0] ^= BIT(pin);
 }
@@ -77,14 +77,14 @@ static inline void gpio_input(int pin) {
                            22, 23, 13, 14, 12, 15, 19, 20, 28, 29,  // 10-19
                            30, 31, 32, 35, 36, 9,  10, 11, 0,  0,   // 20-29
                            0,  0,  7,  8,  5,  6,  1,  2,  3,  4};  // 30-39
-  volatile unsigned long *mux = REG(0X3ff49000);
+  volatile uint32_t *mux = REG(0X3ff49000);
   if (pin < 0 || pin > (int) sizeof(map) || map[pin] == 0) return;
   gpio_output_enable(pin, 0);  // Disable output
   mux[map[pin]] |= BIT(9);     // Enable input
 }
 
 static inline bool gpio_read(int pin) {
-  volatile unsigned long *r = GPIO_IN_REG;
+  volatile uint32_t *r = GPIO_IN_REG;
   if (pin > 31) pin -= 31, r = GPIO_IN1_REG;
   return r[0] & BIT(pin) ? 1 : 0;
 }
